pyproject.toml
[tool.poetry]
name = "slck"
version = "0.1.0"
description = "Simple cli tool to manage your slack workspace"
authors = ["joe-yama <s1r0mqme@gmail.com>"]
license = "MIT"
repository = "https://github.com/joe-yama/slck"

[tool.poetry.scripts]
slck = "slck.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
fire = "^0.4.0"
slack-sdk = ">=3.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pysen = {version = "^0.10.1", extras = ["lint"]}
pytest = "^7.1.1"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy =true
mypy_preset = "strict"
line_length = 88
py_version = "py38"

[[tool.pysen.lint.mypy_targets]]
  paths = [".", "./tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
